refactor: small cleanup
Removed some unnecessary checks, as they weren't needed. This system is designed to run continuously until manually stopped or if a crash occurs.

While this tool is available to the public, I originally created it for a fairly large FiveM server to have easy access to database backups in case of emergencies. If you plan to use this tool, you'll need to adjust it accordingly to fit your needs.